Revert "OAuth state schema (#86)"
This reverts commit a811fb1ad40fed8ebd43b1b9664fa09352b75af5.
